Rename filter.
