+++ bundle/README.md
@@ -27,7 +27,7 @@ Filters include:
 | **GlowFilter**           | [@pixi/filter-glow](https://www.npmjs.com/package/@pixi/filter-glow)                               | [View demo](https://filters.pixijs.download/main/demo/index.html?enabled=GlowFilter)           |
 | **GodrayFilter**         | [@pixi/filter-godray](https://www.npmjs.com/package/@pixi/filter-godray)                           | [View demo](https://filters.pixijs.download/main/demo/index.html?enabled=GodrayFilter)         |
 | **GrayscaleFilter**      | [@pixi/filter-grayscale](https://www.npmjs.com/package/@pixi/filter-grayscale)                     | [View demo](https://filters.pixijs.download/main/demo/index.html?enabled=GrayscaleFilter)      |
-| **HueRotationFilter**    | [@pixi/filter-hue-rotation](https://www.npmjs.com/package/@pixi/filter-hue-rotation)               | [View demo](https://filters.pixijs.download/main/demo/index.html?enabled=HueRotationFilter)    |
+| **HslAdjustmentFilter**    | [@pixi/filter-hsl-adjustment](https://www.npmjs.com/package/@pixi/filter-hsl-adjustment)           | [View demo](https://filters.pixijs.download/main/demo/index.html?enabled=HslAdjustmentFilter)    |
 | **KawaseBlurFilter**     | [@pixi/filter-kawase-blur](https://www.npmjs.com/package/@pixi/filter-kawase-blur)                 | [View demo](https://filters.pixijs.download/main/demo/index.html?enabled=KawaseBlurFilter)     |
 | **MotionBlurFilter**     | [@pixi/filter-motion-blur](https://www.npmjs.com/package/@pixi/filter-motion-blur)                 | [View demo](https://filters.pixijs.download/main/demo/index.html?enabled=MotionBlurFilter)     |
 | **MultiColorFilter**     | [@pixi/filter-multi-color-replace](https://www.npmjs.com/package/@pixi/filter-multi-color-replace) | [View demo](https://filters.pixijs.download/main/demo/index.html?enabled=MultiColorFilter)     |
