+++ filters/hsl-adjustment/src/hsladjustment.frag
@@ -29,7 +29,6 @@ vec3 hueShift(vec3 color, float angle) {
 void main()
 {
     vec4 color = texture2D(uSampler, vTextureCoord);
-
     vec4 result = color;
     float brightness = getBrightness(result);
